refactor: 폴더 분리
