+++ src/sass/_heading.scss
@@ -27,7 +27,7 @@
   font-size: 20px;
   font-weight: 400;
   color: $black;
-  border: #ccc;
+  border: $grey;
 }
 
 .title__input:focus {
