+++ src/ts/chooseSection/openChooseModal.ts
