+++ src/ts/index.ts
@@ -1,6 +1,6 @@
 import chooseSectionRender from './chooseSection/chooseSectionRender';
 import chooseSectionMain from './chooseSection/chooseSectionMain';
-import renderMyFavorite from './renderMyFavorite';
+import renderMyFavorite from './mywaySection/renderMyFavorite';
 import { resetTitle, generateName } from './generateName';
 
 chooseSectionMain();
