+++ src/ts/mywaySection/renderMyFavorite.ts
